Added dashboards auto apply
Fixed closes #671
